Remove bayesian_gmm file
