+++ rdt/transformers/numerical.py
@@ -1,10 +1,13 @@
 """Transformers for numerical data."""
+from collections import namedtuple
 import copy
 import sys
 
 import numpy as np
 import pandas as pd
 import scipy
+from sklearn.mixture import BayesianGaussianMixture
+
 
 from rdt.transformers.base import BaseTransformer
 from rdt.transformers.null import NullTransformer
@@ -489,6 +492,15 @@ class GaussianCopulaTransformer(NumericalTransformer):
         return super()._reverse_transform(data)
 
 
+SpanInfo = namedtuple('SpanInfo', ['dim', 'activation_fn'])
+ColumnTransformInfo = namedtuple(
+    'ColumnTransformInfo', [
+        'column_name', 'column_type',
+        'transform', 'transform_aux',
+        'output_info', 'output_dimensions'
+    ]
+)
+
 class BayesGMMTransformer(NumericalTransformer):
     """Bayesian GMM transformer."""
 
@@ -496,8 +508,10 @@ class BayesGMMTransformer(NumericalTransformer):
     DETERMINISTIC_REVERSE = False
     COMPOSITION_IS_IDENTITY = False
 
-    def __init__(self, dtype=None, nan='mean', null_column=None):
+    def __init__(self, dtype=None, nan='mean', null_column=None, max_clusters=10, weight_threshold=0.005):
         super().__init__(dtype=dtype, nan=nan, null_column=null_column)
+        self._max_clusters = max_clusters
+        self._weight_threshold = weight_threshold
 
     def get_output_types(self):
         """Return the output types supported by the transformer.
@@ -515,17 +529,71 @@ class BayesGMMTransformer(NumericalTransformer):
 
         return self._add_prefix(output_types)
 
-    def _fit(self, data):
+    def _fit_continuous(self, column_name, raw_column_data):
+        """Train Bayesian GMM for continuous column."""
+        gm = BayesianGaussianMixture(
+            n_components=self._max_clusters,
+            weight_concentration_prior_type='dirichlet_process',
+            weight_concentration_prior=0.001,
+            n_init=1
+        )
+
+        gm.fit(raw_column_data.reshape(-1, 1))
+        valid_component_indicator = gm.weights_ > self._weight_threshold
+        num_components = valid_component_indicator.sum()
+
+        return ColumnTransformInfo(
+            column_name=column_name, column_type='continuous', transform=gm,
+            transform_aux=valid_component_indicator,
+            output_info=[SpanInfo(1, 'tanh'), SpanInfo(num_components, 'softmax')],
+            output_dimensions=1 + num_components)
+
+    def _fit(self, data, discrete_columns=()):
         """Fit the transformer to the data.
 
         Args:
             data (pandas.Series):
                 Data to fit to.
         """
-        from rdt.transformers.bayes_gmm import BayesGMMCopy
-        self._model = BayesGMMCopy()
         data = pd.DataFrame(data)
-        self._model._fit(data)
+
+        self.output_info_list = []
+        self.output_dimensions = 0
+        self.dataframe = True
+        self._column_raw_dtypes = data.infer_objects().dtypes
+        self._column_transform_info_list = []
+        for column_name in data.columns:
+            raw_column_data = data[column_name].to_numpy()
+            column_transform_info = self._fit_continuous(column_name, raw_column_data)
+            self.output_info_list.append(column_transform_info.output_info)
+            self.output_dimensions += column_transform_info.output_dimensions
+            self._column_transform_info_list.append(column_transform_info)
+
+    def _transform_continuous(self, column_transform_info, raw_column_data):
+        gm = column_transform_info.transform
+
+        valid_component_indicator = column_transform_info.transform_aux
+        num_components = valid_component_indicator.sum()
+
+        means = gm.means_.reshape((1, self._max_clusters))
+        stds = np.sqrt(gm.covariances_).reshape((1, self._max_clusters))
+        normalized_values = ((raw_column_data - means) / (4 * stds))[:, valid_component_indicator]
+        component_probs = gm.predict_proba(raw_column_data)[:, valid_component_indicator]
+
+        selected_component = np.zeros(len(raw_column_data), dtype='int')
+        for i in range(len(raw_column_data)):
+            component_porb_t = component_probs[i] + 1e-6
+            component_porb_t = component_porb_t / component_porb_t.sum()
+            selected_component[i] = np.random.choice(np.arange(num_components), p=component_porb_t)
+
+        aranged = np.arange(len(raw_column_data))
+        selected_normalized_value = normalized_values[aranged, selected_component].reshape([-1, 1])
+        selected_normalized_value = np.clip(selected_normalized_value, -.99, .99)
+
+        selected_component_onehot = np.zeros_like(component_probs)
+        selected_component_onehot[np.arange(len(raw_column_data)), selected_component] = 1
+
+        return [selected_normalized_value, selected_component_onehot]
 
     def _transform(self, data):
         """Transform numerical data.
@@ -538,15 +606,46 @@ class BayesGMMTransformer(NumericalTransformer):
             numpy.ndarray
         """
         data = pd.DataFrame(data)
-        result = self._model._transform(data)
-        normalized, one_hot = result[:, 0], result[:, 1:]
-        one_hot_as_label = one_hot.argmax(axis=1)  # something like this?
+        column_data_list = []
+        for column_transform_info in self._column_transform_info_list:
+            column_data = data[[column_transform_info.column_name]].to_numpy()
+            if column_transform_info.column_type == 'continuous':
+                column_data_list += self._transform_continuous(column_transform_info, column_data)
+
+        transformed = np.concatenate(column_data_list, axis=1).astype(float)
+        normalized, one_hot = transformed[:, 0], transformed[:, 1:]
+        one_hot_as_label = one_hot.argmax(axis=1)
         return pd.DataFrame({
             'continuous': normalized,
             'discrete': one_hot_as_label
         })
 
-    def _reverse_transform(self, data):
+    def _reverse_transform_continuous(self, column_transform_info, column_data, sigmas, st):
+        gm = column_transform_info.transform
+        valid_component_indicator = column_transform_info.transform_aux
+
+        selected_normalized_value = column_data[:, 0]
+        selected_component_probs = column_data[:, 1:]
+
+        if sigmas is not None:
+            sig = sigmas[st]
+            selected_normalized_value = np.random.normal(selected_normalized_value, sig)
+
+        selected_normalized_value = np.clip(selected_normalized_value, -1, 1)
+        component_probs = np.ones((len(column_data), self._max_clusters)) * -100
+        component_probs[:, valid_component_indicator] = selected_component_probs
+
+        means = gm.means_.reshape([-1])
+        stds = np.sqrt(gm.covariances_).reshape([-1])
+        selected_component = np.argmax(component_probs, axis=1)
+
+        std_t = stds[selected_component]
+        mean_t = means[selected_component]
+        column = selected_normalized_value * 4 * std_t + mean_t
+
+        return column
+
+    def _reverse_transform(self, data, sigmas=None):
         """Convert data back into the original format.
 
         Args:
@@ -559,4 +658,26 @@ class BayesGMMTransformer(NumericalTransformer):
         data = pd.DataFrame(data)
         one_hot = pd.get_dummies(data['a.discrete'])
         data = np.concatenate([data['a.continuous'][:, None], one_hot], axis=1)
-        return self._model._reverse_transform(data)
+        
+        st = 0
+        recovered_column_data_list = []
+        column_names = []
+        for column_transform_info in self._column_transform_info_list:
+            dim = column_transform_info.output_dimensions
+            column_data = data[:, st:st + dim]
+            if column_transform_info.column_type == 'continuous':
+                recovered_column_data = self._reverse_transform_continuous(
+                    column_transform_info, column_data, sigmas, st)
+
+            recovered_column_data_list.append(recovered_column_data)
+            column_names.append(column_transform_info.column_name)
+            st += dim
+
+        recovered_data = np.column_stack(recovered_column_data_list)
+        recovered_data = (pd.DataFrame(recovered_data, columns=column_names)
+                          .astype(self._column_raw_dtypes))
+        
+        if not self.dataframe:
+            recovered_data = recovered_data.to_numpy()
+
+        return recovered_data
