+++ tests/contributing.py
@@ -9,9 +9,9 @@ import coverage
 import numpy as np
 import pandas as pd
 import pytest
+from tabulate import tabulate
 
 from rdt.transformers import get_transformer_class, get_transformers_by_type
-from tabulate import tabulate
 from tests.code_style import (
     get_test_location, validate_test_location, validate_test_names, validate_transformer_addon,
     validate_transformer_importable_from_parent_module, validate_transformer_module,
